Changed from follower to MotorControllerGroup(s)
We've got the robot to move, but only one motor on one side. I'm hoping that with a re-wiring of the motorcontrollers (I don't know how to wire CAN) and a change from followers to controllergroups (the follower controllers never activated when the robot was enabled) I'm hoping that this is going to work.